# openapi.yaml
openapi: 3.0.0
info:
  title: Daily Journal API
  version: 1.0.0
  description: A simple API for managing daily journal entries and users with authentication
servers:
  - url: https://cse341-rlcp.onrender.com
    description: Production server
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3001
    description: Alternative local server (port 3001)
  - url: http://localhost:3002
    description: Alternative local server (port 3002)
  - url: http://localhost:3003
    description: Alternative local server (port 3003)
  - url: http://localhost:3004
    description: Alternative local server (port 3004)
  - url: http://localhost:3005
    description: Alternative local server (port 3005)
components:
  schemas:
...
    User:
      type: object
      required:
        - username
        - email
      properties:
        _id:
          type: string
          description: The auto-generated id of the user
        username:
          type: string
          description: The username of the user
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          description: The email address of the user
        password:
          type: string
          description: The password of the user (min 6 characters)
          minLength: 6
        googleId:
          type: string
          description: Google OAuth ID for the user
        githubId:
          type: string
          description: GitHub OAuth ID for the user
        role:
          type: string
          enum: [user, admin]
          default: user
          description: The role of the user
        isActive:
          type: boolean
          default: true
          description: Whether the user account is active
        lastLogin:
          type: string
          format: date-time
          description: The last login timestamp
        preferences:
          type: object
          properties:
            theme:
              type: string
              enum: [light, dark]
              default: light
            notifications:
              type: boolean
              default: true
            language:
              type: string
              default: en
        createdAt:
          type: string
          format: date-time
          description: The creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: The last update timestamp
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    githubAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            user:email: Read user email
            read:user: Read user profile
paths:
...
  /api/users:
    get:
      summary: Returns all users (requires authentication)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    post:
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
        '500':
          description: Server error
  /api/users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: User ID

    get:
      summary: Returns a user by ID (requires authentication)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A single user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    put:
      summary: Update a user by ID (requires authentication)
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error

    delete:
      summary: Delete a user by ID (requires authentication)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error 